Restructured project so header and code files are not populating the root directory of the repo.
