+++ Makefile
@@ -1,10 +1,12 @@
 BIN=./dss
+INCLUDE=./include
+SRC=./src
 
-all: tree dss.cu
-	$(CC) kd_tree.o dss.cu -o $(BIN)
+all: tree $(SRC)/dss.cu
+	$(CC) kd_tree.o $(SRC)/dss.cu -o $(BIN)
 
-tree: kd_tree.h kd_tree.c
-	$(CC) -c kd_tree.c
+tree: $(INCLUDE)/kd_tree.h $(SRC)/kd_tree.c
+	$(CC) -c $(SRC)/kd_tree.c
 
 clean:
 	rm -rf *.o $(BIN)
